Fix customer ID/version number getting cut off

+++ software/slots.h
@@ -3,8 +3,8 @@
 
 #define N_SLOTS 8
 
-int read_slot_data(int slot, long *data);
-void write_slot_data(int slot, long data);
+int read_slot_data(int slot, unsigned long long *data);
+void write_slot_data(int slot, unsigned long long data);
 void clear_slot_data(int slot);
 
 #endif // SLOTS_H
